# web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Swipe Shopify"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-access-token"
            }
        }
    },
    "security": [
        {
            "apiKeyAuth": []
        }
    ],
    "paths": {
...
        "/merchant/setting/upsert": {
            "post": {
                "tags": ["Merchant"],
                "summary": "Setting upsert API",
                "requestBody": {
                    "description": "Setting upsert Payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "billing_contact_email": {
                                        "type": "string",
                                        "example": "post-protection.myshopify.com"
                                    },
                                    "reimbursement_contact_email": {
                                        "type": "string",
                                        "example": "post-protection.myshopify.com"
                                    },
                                    "claims_contact_email": {
                                        "type": "string",
                                        "example": "post-protection.myshopify.com"
                                    },
                                    "route_plus_package_protection": {
                                        "type": "string",
                                        "example": "on"
                                    },
                                    "protection_widget": {
                                        "type": "string",
                                        "example": "on"
                                    },
                                    "thank_you_page_tracking_link": {
                                        "type": "string",
                                        "example": "on"
                                    },
                                    "free_shipping_and_protection_bar": {
                                        "type": "string",
                                        "example": "on"
                                    }
                                }
                            }
                        }
                    }
                },
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "data Update Successfully!"
                    },
                    "400": {
                        "description": "Merchant does not exist."
                    }
                }
            }
        },
...
        "/merchant/setting/save-account": {
            "post": {
                "tags": ["Setting"],
                "summary": "Save Merchant Account Setting API",
                "requestBody": {
                    "description": "Merchant Account Setting Payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Save-Account"
                            }
                        }
                    }
                },
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Comment added successfully"
                    },
                    "400": {
                        "description": "Comment did not added"
                    }
                }
            }
        },
        "/merchant/setting/save-bank": {
            "post": {
                "tags": ["Setting"],
                "summary": "Save Merchant Bank Account Details API",
                "requestBody": {
                    "description": "Merchant Bank Account Details Payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Save-Bank"
                            }
                        }
                    }
                },
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Comment added successfully"
                    },
                    "400": {
                        "description": "Comment did not added"
                    }
                }
            }
        },
        "/merchant/setting/get-account/{type}": {
            "get": {
                "tags": ["Setting"],
                "summary": "Merchant Account Details API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data found"
                    },
                    "400": {
                        "description": "No Records found"
                    }
                }
            }
        },
        "/merchant/setting/retrieve-payment": {
            "get": {
                "tags": ["Setting"],
                "summary": "Retrive Merchant Payment Details API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Data found"
                    },
                    "400": {
                        "description": "No Records found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Save-Account":{
            "type": "object",
            "properties": {
                "payment_method ": {
                    "type": "string",
                    "example": "pm_1OHiGGDcgoMRnDdZUnBThRMJ"
                }
            }
        },
        "Save-Bank":{
            "type": "object",
            "properties": {
                "bank_token_id": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
    }
}